refactor: 디렉토리 구조 리팩토링
- 각 페이지별 함수를 한 파일에 모으는 것으로 변경
